Correctly show parsed block message in Suggested Edits screen.
