+++ app/src/main/java/org/wikipedia/dataclient/mwapi/UserInfo.java
@@ -14,17 +14,12 @@ import java.util.List;
 import java.util.Set;
 
 @SuppressWarnings("unused")
-public class UserInfo {
+public class UserInfo extends MwServiceError.BlockInfo {
     private String name;
     private int id;
     @Nullable private List<String> groups;
-    private int blockid;
     private int editcount;
     @Nullable private String latestcontrib;
-    @Nullable private String blockreason;
-    @Nullable private String blockedby;
-    @Nullable private String blockedtimestamp;
-    @Nullable private String blockexpiry;
 
     public int id() {
         return id;
@@ -44,14 +39,5 @@ public class UserInfo {
             date = DateUtil.iso8601DateParse(latestcontrib);
         }
         return date;
-    }
-
-    public boolean isBlocked() {
-        if (TextUtils.isEmpty(blockexpiry)) {
-            return false;
-        }
-        Date now = new Date();
-        Date expiry = DateUtil.iso8601DateParse(blockexpiry);
-        return expiry.after(now);
     }
 }
